fix(zen-cloud-vastai): f64 q_grid — missed caller from sweep f64-q change
The 759ab501 sweep f64-q change (parse_q_grid -> Vec<f64>) missed a caller in
zen-cloud-vastai/src/worker/sweep_runner.rs:169 (`let q_grid: Vec<u32>`),
which only surfaces building zen-sweep-worker --features vastai (the zen-metrics-cli
CPU lib test didn't cover it). q_grid flows straight into SweepConfig{q_grid}
(now Vec<f64>), so the annotation is the whole fix. Restores the fleet-worker
build.

+++ crates/zen-cloud-vastai/src/worker/sweep_runner.rs
@@ -166,7 +166,7 @@ pub fn run_group_inline(spec: InlineGroupSpec) -> Result<()> {
     // isn't Send+Sync — anyhow's Context trait won't bind to it
     // directly. Convert via the error message instead; the caller
     // doesn't need the full error chain.
-    let q_grid: Vec<u32> = parse_q_grid(&spec.q_grid)
+    let q_grid: Vec<f64> = parse_q_grid(&spec.q_grid)
         .map_err(|e| anyhow::anyhow!("parse q_grid {:?}: {e}", spec.q_grid))?;
     let knob_grid: KnobGrid = parse_knob_grid(&spec.knob_grid_json)
         .map_err(|e| anyhow::anyhow!("parse knob_grid {:?}: {e}", spec.knob_grid_json))?;
